fix: removed unwanted components for signin & signup page
