+++ apps/retail/pages/signin.tsx
@@ -1,16 +1,31 @@
-import { Box } from '@chakra-ui/react'
 import React, { useEffect } from 'react'
-import Signin from '@components/signIn/SignIn'
+import Router from 'next/router'
+import kuzaMobLogo from '@public/images/Logo.svg'
+import kuzaDeskLogo from '@public/images/KuzaLogo.svg'
+import { useLanguage } from '@hooks/useLanguage'
+import { SignInPage } from '@beckn-ui/common'
 
 const Login = () => {
+  const { t } = useLanguage()
+
   useEffect(() => {
     localStorage.clear()
   }, [])
 
+  const handleSignUp = () => {
+    Router.push('/signUp')
+  }
+
   return (
-    <Box>
-      <Signin />
-    </Box>
+    <SignInPage
+      logos={{
+        mobile: { src: kuzaMobLogo, alt: 'Kuza logo' },
+        desktop: { src: kuzaDeskLogo, alt: 'Kuza logo' }
+      }}
+      onSignIn={() => {}}
+      onSignUp={handleSignUp}
+      t={key => t[key]}
+    />
   )
 }
